% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum_limit_exact.R
\name{calc_cusum}
\alias{calc_cusum}
\title{Calculate CUSUM}
\usage{
calc_cusum(x, c0, cA)
}
\arguments{
\item{x}{vector of outcomes}

\item{c0}{accepted failure probability}

\item{cA}{smallest detectable failure probability}
}
\value{
Returns matrix of possible sequences
}
\description{
This function calculates the CUSUM chart for the given sequence of successes and failures
}
