% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gcusum}
\alias{gcusum}
\title{Grouped-CUSUM chart}
\usage{
gcusum(input_outcomes, failure_prob, delta, control_limit,
  max_num_shuffles, seed, quantiles)
}
\arguments{
\item{input_outcomes}{NumericMatrix, patient outcomes and block_id (continuous).}

\item{failure_prob}{double, accepted failure probability of null hypothesis.}

\item{delta}{double, odds multiplier.}

\item{control_limit}{double, control limit.}

\item{max_num_shuffles}{integer, number of shuffles.}

\item{seed}{integer.}

\item{quantiles}{NumericVector, quantiles that are returned.}
}
\value{
gcusum NumericMatix, signal probability, average CUSUM value and specified quantiles for every observation.
}
\description{
Calculate GCUSUM chart for non-risk-adjusted processes.
}
