% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum_limit_exact.R
\name{make_all_outcomes}
\alias{make_all_outcomes}
\title{Make all outcomes}
\usage{
make_all_outcomes(npat_outcome)
}
\arguments{
\item{npat_outcome}{Number of patients (sample sizes)}
}
\value{
Returns matrix of possible sequences
}
\description{
creates all possible sequences of outcomes for a sample size
}
