% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cusum.R
\name{plot.cusum}
\alias{plot.cusum}
\title{Plot CUSUM chart for a cusum Object}
\usage{
## S3 method for class 'cusum' plot(x, signal = TRUE, ...)
}
\arguments{
\item{x}{An object of class cusum}

\item{signal}{Logical. If TRUE, signals are plotted (default)}
}
\description{
Produces a CUSUM chart.
}
