% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ragcusum}
\alias{ragcusum}
\title{RA-Grouped-CUSUM chart}
\usage{
ragcusum(input_ra_outcomes, control_limit, max_num_shuffles, seed,
  quantiles)
}
\arguments{
\item{input_ra_outcomes}{NumericMatrix, patient outcomes and block_id (continuous).}

\item{control_limit}{double, control limit.}

\item{max_num_shuffles}{integer, number of shuffles.}

\item{seed}{integer.}

\item{quantiles}{NumericVector, quantiles that are returned.}
}
\value{
ragcusum NumericMatix, signal probability, average CUSUM value and specified quantiles for every observation.
}
\description{
Calculate GCUSUM chart for risk-adjusted processes.
}
