% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ragscusum_example_data}
\alias{ragscusum_example_data}
\title{Group-sequential Risk-adjusted Performance Data with Block Identifier}
\format{A data frame with 2000 rows and 4 variables:
\describe{
  \item{t}{Sequence of observations}
  \item{y}{Patient outcome}
  \item{score}{Patient risk score}
  \item{year}{Year of treatment}
  \item{block_identifier}{Continuous block identifier}
}}
\source{
Data for simulation was provided by Bavarian Agency of Quality Assurance (BAQ), Munich Germany.

Description of performance indicator (in German): \url{https://iqtig.org/downloads/auswertung/2016/10n2karot/QSKH_10n2-KAROT_2016_QIDB_V02_2017-04-26.pdf}
}
\usage{
ragscusum_example_data
}
\description{
Generated performance data of indicator: Ratio of observed to expected cases of severe stroke or death under open carotid stenosis surgery.
}
\details{
Individual patient risk scores were drawn from actual hospital data and patient outcomes were simulated. Two years are provided, so Phase I and Phase II can be defined.
}
\keyword{datasets}
