% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cusum.R
\name{summary.cusum}
\alias{summary.cusum}
\title{summary for a cusum Object}
\usage{
## S3 method for class 'cusum' summary(x, ...)
}
\arguments{
\item{x}{An object of class cusum}
}
\description{
summary cusum results
}
