% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_t.R
\name{weights_t}
\alias{weights_t}
\title{Weights for observations}
\usage{
weights_t(patient_outcomes, probability_ae, odds_multiplier = 2)
}
\arguments{
\item{patient_outcomes}{Integer. Vector of binary patient outcomes (0,1)}

\item{probability_ae}{Double. Baseline failure probability for adverse event in non-risk-adjusted case, vector of patient risk scores for risk-adjustment.}

\item{odds_multiplier}{Double. Odds multiplier of adverse event under the alternative hypothesis (<1 looks for decreases)}
}
\description{
Calculate standard CUSUM weights
}
