% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cusum_example_data}
\alias{cusum_example_data}
\title{Non-Risk-adjusted Performance Data}
\format{A data frame with 33270 rows and 3 variables:
\describe{
  \item{t}{Temporal sequence of observations}
  \item{y}{Patient outcome}
  \item{year}{Year of treatment}
}}
\source{
Decription of performance indicator (in German): \url{https://iqtig.org/downloads/auswertung/2016/17n1hftfrak/QSKH_17n1-HUEFTFRAK_2016_QIDB_V02_2017-04-26.pdf}
}
\usage{
cusum_example_data
}
\description{
Performance data of indicator 54030: Preoperative stay more than 24 hours for patients with proximal femur fracture for Bavarian hospitals in the years 2016-2017.
}
\keyword{datasets}
