% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gcusum}
\alias{gcusum}
\title{Grouped-CUSUM chart}
\usage{
gcusum(input_outcomes, failure_probability, odds_multiplier, limit,
  max_num_shuffles, quantiles, seed = 0L)
}
\arguments{
\item{input_outcomes}{Matrix. First column are binary patient outcomes (0,1). Second column are continuous sequence of block identifier.}

\item{failure_probability}{Double. Baseline failure probability}

\item{odds_multiplier}{Double. Odds multiplier of adverse event under the alternative hypothesis (<1 looks for decreases)}

\item{limit}{Double. Control limit for signalling performance change}

\item{max_num_shuffles}{Integer. Number of shuffles (i.e. different sequences of observations)}

\item{quantiles}{Double. Vector of requested quantiles of GCUSUM distribution}

\item{seed}{Integer. Seed for RNG}
}
\value{
gcusum matrix, signal probability, average CUSUM value and specified quantiles for every observation.
}
\description{
Calculate GCUSUM chart for non-risk-adjusted processes.
}
