% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{racusum_example_data}
\alias{racusum_example_data}
\title{Risk-adjusted Performance Data}
\format{A data frame with 12018 rows and 4 variables:
\describe{
  \item{t}{Temporal sequence of observations}
  \item{y}{Patient outcome}
  \item{score}{Patient risk score}
  \item{year}{Year of treatment}
}}
\source{
Decription of performance indicator (in German): \url{https://iqtig.org/downloads/auswertung/2016/10n2karot/QSKH_10n2-KAROT_2016_QIDB_V02_2017-04-26.pdf}
}
\usage{
racusum_example_data
}
\description{
Performance data of indicator: Ratio of observed to expected cases of severe stroke or death under open carotid stenosis surgery for Bavarian hospitals in the years 2016-2017.
}
\keyword{datasets}
