% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racusum_limit_dpcl.R
\name{racusum_limit_dpcl}
\alias{racusum_limit_dpcl}
\title{Dynamic Probability Control Limits (DPCL)}
\usage{
racusum_limit_dpcl(patient_risks, N = 1e+05, odds_multiplier = 2,
  alpha, seed = NULL)
}
\arguments{
\item{patient_risks}{A vector containing patient risk scores}

\item{N}{Number of simulations}

\item{odds_multiplier}{Odds multiplier for the alternative hypothesis (<1 looks for decreases)}

\item{alpha}{False signal probability}

\item{seed}{An optional seed for simulation}
}
\value{
Returns the control limit
}
\description{
Set DPCL for risk-adjusted Bernoulli CUSUM Charts
}
\examples{
patient_risks <- runif(100, min = 0.1, max = 0.8)

racusum_limit_dpcl(patient_risks = patient_risks,
N = 1000,
odds_multiplier = 2,
alpha = 0.05,
seed = 32423)


}
\references{
Zhang, Xiang & Woodall, William. (2016). Dynamic Probability Control Limits for Lower and Two-Sided Risk-Adjusted Bernoulli CUSUM Charts. Quality and Reliability Engineering International. 10.1002/qre.2044.
}
